Updated to handle multiple invocations of motion(1) command to allow more than 10 cameras webcontrol ports start at specified (e.g. 8080) stream_ports start at specified (e.g. 9090); configured to handle 30 cameras tested with 20 added controls for file:// specification of directory receiving either jpg or 3gp files jpg files are copied into latest.jpg in order received 3gp files are bbroken into jpg and submitted in order with delay according to FPS specified.

+++ motion/on_event_end.sh
@@ -399,14 +399,14 @@ if ($?MOTION_MQTT_HOST && $?MOTION_MQTT_PORT) then
   if ($?VERBOSE) mosquitto_pub -h "${MOTION_MQTT_HOST}" -t "debug" -m '{"VERBOSE":"'$0:t'","pid":'$$',"topic":"'"$MQTT_TOPIC"'"}'
 endif
 
-## cleanup
-rm -fr $tmpdir
-
 ##
 ## ALL DONE
 ##
 
 done:
+  if ($?tmpdir) then
+    rm -fr $tmpdir
+  endif
   echo "$0:t $$ -- END" `date` >& /dev/stderr
   if ($?VERBOSE) mosquitto_pub -h "${MOTION_MQTT_HOST}" -t "debug" -m '{"VERBOSE":"'$0:t'","pid":'$$',"info":"END"}'
   exit
